# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HIX-AI-2API",
        "title": "HIX-AI-2API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/key": {
            "put": {
                "description": "保存API-KEY",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API-KEY"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "API-KEY信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApiKeySaveReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/key/all": {
            "get": {
                "description": "查询全量API-KEY",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API-KEY"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ApiKeyResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/key/update": {
            "post": {
                "description": "更新API-KEY",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API-KEY"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "API-KEY信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApiKeyUpdateReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/key/{id}": {
            "delete": {
                "description": "删除API-KEY",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API-KEY"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API-KEY ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "common.ResponseResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ApiKeyResp": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "description": "ApiKey apiKey",
                    "type": "string"
                },
                "createTime": {
                    "description": "CreateTime 创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "Id Id",
                    "type": "string"
                },
                "remark": {
                    "description": "Remark 备注",
                    "type": "string"
                }
            }
        },
        "model.ApiKeySaveReq": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "description": "ApiKey apiKey",
                    "type": "string"
                },
                "remark": {
                    "description": "Remark 备注",
                    "type": "string"
                }
            }
        },
        "model.ApiKeyUpdateReq": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "description": "ApiKey apiKey",
                    "type": "string"
                },
                "id": {
                    "description": "Id id",
                    "type": "string"
                },
                "remark": {
                    "description": "Remark 备注",
                    "type": "string"
                }
            }
        },
...
    }
}